kathulis-backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
